Feat/chooseinputtype (#2)
* Initial Attempt At Functionality

The basic idea is here, but there are a few bugs: The radio buttons aren't exclusive (which should be relatively simple to fix) and the one line handler crashes when not passed a bonus to add.

* Bug Fixes

Fixed the bug where not adding a bonus threw an uncaught error, the (previously unnoticed) display bug regarding bonus, and the radio button exclusivity.

* Updated Gitignore

An update to the gitignore to prepare for when I import the project into eclipse

* Minor improvement

With this minor improvement in functioning of onelinedice and overall setup of code, this branch is ready for dev. Next suggested feature is more descriptive error handling, and overall code reorganization.