Feat/cleanup (#3)
* Initial attempt at cleanup

Divided all the handlers into separate classes. Unsure if it works
properly still, I have yet to test it. However, I need to commit all the
changes that I've made so far.

* Post-Testing Changes

After a little bit of testing, it was shown that the original method
doesn't necessarily work. However, rather than attempting to deal with
producing and firing custom events to store the data I need, I simply
update the action events via the previously created constructors and the
built-in ability to listen for changes in the various inputs. Hopefully
the previous version of the input handlers are discarded on creation of
the new one, or one could run into memory issues running the program too
long.