Image Support (Part 2)
d100 are now supported. I'm honestly considering moving the whole result-stage-making out of utility, since that makes the class a whopping 780 lines long, and that's a bit unweildy. However, that'd be a change to take place after merging into dev. I think I'm ready to close the book on the dice again, after we get all the merges done.