In Package Images
So after some tinkering, keeping the images in the project structure actually winds up being rather troublesome, when we package the project as a jar. While this is salvagable, I'm going to have to host the images somewhere online to fix it, and I'm off work now. Will tinker further tomorrow