Feat/stream work (#4)
* Improved Error Handling

This is just an initial commit, the error messages are more friendly to me, the programmer, but they don't make much sense to the user. So, we'll be fixing that.

* User-Readable Error Messages

Some minor changes to the way errors are handled, in order to make the user causing problems at least a little less intimidated.

* Window Management

Just some small changes to the showCorrectInputHandler to keep things clean, so there aren't 3 whole windows on screen all the darn time.

* Scrolled Bar

Added a scrollbar, as requested by the beta testers. They were worried about not being able to see all their dice!

* File Logging

Just included the standard file-based logging that I've included in a few other projects, in case someone manages to find some funky bug that I haven't seen yet.

* Actually use the file based logging

Actually use the file based logging we implemented. Whoops, almost forgot to use it!

* Fix File Output Troubles

The file outputs (both for errors and for ordinary outputs) were having some trouble due to a minor error, it's now corrected! Hooray!

* Attempts at Images

Doesn't display the images yet, but this code should be the framework to show images for four sided dice, and the rest of the dice images should follow a similar idea. Also, enums, yay!

* Image Support (part 1)

The currently used paths are just test paths, but the dice now support images for all common dice types except for 1d100, as the rolling of percentile dice is more complex. Thankfully, the way I've implemented the enums, it should be simple to actually update these paths from their testing state. Additionally, I'm not doing the remainder of the work on stream, I don't want to burn out coding all day at work, then the whole stream.

* Image Support (Part 2)

d100 are now supported. I'm honestly considering moving the whole result-stage-making out of utility, since that makes the class a whopping 780 lines long, and that's a bit unweildy. However, that'd be a change to take place after merging into dev. I think I'm ready to close the book on the dice again, after we get all the merges done.

* In Package Images

So after some tinkering, keeping the images in the project structure actually winds up being rather troublesome, when we package the project as a jar. While this is salvagable, I'm going to have to host the images somewhere online to fix it, and I'm off work now. Will tinker further tomorrow.

* Images Proper

Took some tinkering, but not only did I remove 200+ extraneous lines of code, but I also made it so that the images are able to be packaged into the full program now. With these changes, this branch is ready for dev.